Schedule saving and appointment data entry creation
