+++ UniSA Radiation Therapy Mock Clinic Scheduler/Views/Coordinator/CreateClinic.cshtml
@@ -9,6 +9,11 @@
 <script src="https://code.jquery.com/jquery-1.11.0.min.js"></script>
 <script src="https://netdna.bootstrapcdn.com/bootstrap/3.1.1/js/bootstrap.min.js"></script>
 <script src="~/js/moment.min.js"></script>
+
+<script src="~/js/classes/TableManager.js"></script>
+<script src="~/js/classes/AJAXManager.js"></script>
+<script src="~/js/classes/ScheduleManager.js"></script>
+<script src="~/js/classes/FormWizard.js"></script>
 <script src="~/js/createClinic.js"></script>
 
 <div class="container">
