+++ UniSA Radiation Therapy Mock Clinic Scheduler/Views/Shared/_Layout.cshtml
@@ -107,7 +107,7 @@
                                 <div class="list-group-item @((string?) ViewData["Title"] == "Clinics" ? "active" : "")">
                                     <div class="list-group-item-inner-container">
                                         <a class="list-group-item-action list-indent-1" asp-controller="Home" asp-action="Clinics">
-                                            <span class="nav-item-text">@(coordinator ? "My Clinics" : "My Appointments")</span>
+                                            <span class="nav-item-text">@(coordinator ? "Upcoming Clinics" : "Appointments")</span>
                                         </a>
                                     </div>
                                 </div>
